requirements.txt
opencv-python==4.10.0.84
numpy==2.2.1
pandas==2.2.3
pillow==11.1.0
natsort==8.4.0
requests==2.32.3



certifi==2022.12.7
charset-normalizer==2.1.1
filelock==3.13.1
fsspec==2024.6.1
idna==3.4
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3



sympy==1.13.3
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
typing_extensions==4.12.2
urllib3==1.26.13
